server/services/linkedin-scraper/requirements.txt
numpy==1.24.3
flask==2.0.1
flask-cors==4.0.0
pandas==2.0.3
playwright==1.20.0
openai==0.27.0
python-dotenv==0.19.0
Werkzeug==2.0.1
python-docx==0.8.11
pdfminer.six==20221105
beautifulsoup4==4.11.1
lxml==4.9.2
gunicorn==20.1.0
requests==2.31.0
